Implemented aggregate.Changed
The Aggregate.Changed message will contain the data of any change that happens to a aggregate root.
This Message will wrap any domain events (source data) that need to be applied to an aggregate root and contain the (processing data) of the domain event.

The message will be used the record changes and reconstruct the state of a aggregate root by reconstructing the message.