Implement aggregate.Root
The aggregate.Root is a interface that a applications aggregate root must implement.
In order to implement this interface the applications aggregate root must embed the `aggregate.BaseRoot`.

The `aggregate.BaseRoot` is responsible for recording and providing the recorded `aggregate.Changed` messages.

The application is responsible for implementing the `AggregateID` and `Apply` (see `aggregate.EventApplied`) functions of the AggregateRoot.
The `AggregateID` function return the identifier of the aggregate root and must be set before any events can be applied.
The `Apply` function is used to update the current state of the AggregateRoot and it's related Aggregates.