Implemented sql.MessageFactory
In order to reconstruct messages the an SQL database we need a way to create them.
The Message Factory is used to transform the database rows into messages.

In order handle aggregate.Changed message the sql.AggregateChangedFactory has been introduced.
This should support recreation of aggregate.Changed messages and there Payloads.