Test metadata Unmarshall with AsMap function
Since golang maps will not ensure keys order we cannot enforce the metadata to contain the data in this order.
For now we will fix test by transforming the Metadata to a Map in the future (if the json lexer ever becomes public) this should be done using a custom unmarshal.