Introducing the projection interfaces
The projection interfaces set the base for projections will work in goengine.
The projection logic is split in 2 part the first being the projection which can be viewed as the configuration and
the projector that will execute the projection that will manage the state of the projection in regard to an event stream.