Introducing the postgres.ReadOnlyEventStore
When the projector lost connection and reconnected the event store was using another connection pool which caused a ErrBadConn.  To avoid this error the event store can now use a provided connection.

This change allows the projector and it's related process to use a single connection pool instead of 2 separate once.