Extract shared integration tests code
Since a lot of code regarding integration tests for projectors will be shared it's time to extract it into a nice little suite that all of them can use.
This include the setup of the eventStore and the appending of events.

As a added bonus we also moved the domain used for the integration tests into the same file.