Projections may be marked failed due to early return of pg_advisory_unlock
In some cases when notifications are queued if may happen that the advisory lock is released before before the Update statement actually set the locked column to false.
To avoid this issue the query needed to be split into 2 queries.

It was also tried to use a WITH but the had the same result as before.