Improve metadata unmarshal performance
While profiling a application that uses goengine. The amount of memory used for Unmarshalling metadata was extreme.
By using the json.NewDecoder instead of unmarshalling to a `map[string]interface{}` the speed and memory usage has gone down.

Before this PR the add benchmark was:
```
// BenchmarkJSONMetadata_UnmarshalJSON-4   	  300000	      4313 ns/op	    1208 B/op	      29 allocs/op
```
With the new and improved version it has become:
```
BenchmarkJSONMetadata_UnmarshalJSON-4   	2000000000	         0.00 ns/op	       0 B/op	       0 allocs/op
```